Fix C4267: cast size_t/unsigned long to int32_t at narrow call sites
- basisset_map.hpp: cast size_t st_idx/range_end to int32_t when inserting
  into vector<int32_t> and vector<pair<int32_t,int32_t>>
- xc_task.hpp: use size_t{0} init in std::accumulate (was 0ul, which is
  32-bit on Windows)
- load_balancer_impl.cxx: same 0ul -> size_t{0} fix in total_npts()
- molmeta.cxx: same 0ul -> size_t{0} fix in sum_atomic_charges accumulate
- exx_screening.cxx: cast shell index loop vars to int32_t when pushing
  into vector<pair<int32_t,int32_t>> and vector<int32_t>

+++ include/gauxc/basisset_map.hpp
@@ -61,8 +61,8 @@ public:
     size_t st_idx = 0;
     for( const auto& shell : basis ) {
       size_t range_end = st_idx + shell.size();
-      shell_to_first_ao_.emplace_back( st_idx );
-      shell_to_ao_range_.push_back({ st_idx, range_end });
+      shell_to_first_ao_.emplace_back( static_cast<int32_t>(st_idx) );
+      shell_to_ao_range_.push_back({ static_cast<int32_t>(st_idx), static_cast<int32_t>(range_end) });
       st_idx = range_end;
     }
